refactor: Apply CuratedArtWorkAttribute Name
