+++ src/components/generate/CuratedArtworkSection/CuratedArtworkForm.tsx
@@ -1,8 +1,8 @@
-import { _CuratedContentType } from '@/types/curatedArtwork-types';
 import React, { useReducer } from 'react';
 import CustomInput from '../../CustomInput';
 import DropdownMenu from '../../DropdownMenu';
 import { FormState, getEmptyFormState } from './states';
+import { _CuratedContentType } from '@/model/interface/curatedArtwork-types';
 
 interface CuratedArtworkFormProps {
     addCuratedArtwork: (formState: FormState) => void;
