+++ src/model/interface/curatedArtwork-types.ts
@@ -9,7 +9,7 @@ export const curatedContentType = {
 export type _CuratedContentType = keyof typeof curatedContentType;
 
 // TODO CuratedWorkAttribute이름 변경 CuratedArtWorkAttribute이름
-export interface CuratedWorkAttribute {
+export interface CuratedArtWorkAttribute {
     id: string; // TODO ID는 Painting ID를 따라갈지 고민
     type: _CuratedContentType;
     cldId: string;
